Benchmarks added but not enforced.