@michaelficarra 2016-11-11
You should probably move the benchmarks out of the tests into something with a `main` method. If they don't make assertions, they aren't tests.