@ikarienator 2016-11-11
Can I actually start to assert on the performance numbers?