@michaelficarra 2016-11-11
I don't think we should make perf assertions. I just think the benchmarking files should not be marked as tests. This will also fix the timeout issue. We don't need to run benchmarks every time we run tests, especially if there's no assertions.